refactor(review): 코드래빗 리뷰 반영 수정 (#72)
