+++ apps/homepage/src/constants/mypage-mem/term/PrivacyPolicyData.tsx
